refactor: Racing 객체 도메인 분류
Racing 속 UI로 판단되는 함수(start())를 printMessage.js로 이동(이후 refactor 필요)
우승자를 출력할 때 바로 출력형식으로 바꾸는 게 아니라 배열로 반환하도록 변경
불필요한 count 필드 삭제

이를 통해 Racing 객체가 도메인 역할만 하도록 기능을 분류함
