+++ __tests__/Racing.test.js
@@ -1,21 +1,22 @@
-import Racing from '../src/Racing.js';
-import Car from '../src/Car.js';
+import Racing from '../src/domain/Racing.js';
+import Car from '../src/domain/Car.js';
 
 describe('자동차 경주 객체 테스트', () => {
   const carNames = ['abc', 'def', 'efg'];
   const carList = carNames.map((carName, i) => new Car(carName, i));
-  const count = 5;
-  const racing = new Racing(carList, count);
+  const racing = new Racing(carList);
 
   test('자동차 객체 정보를 담는다.', () => {
-    expect(racing.carList.map((car) => car.name)).toEqual(carNames);
+    expect(racing.carList.map((car) => car.getCarStatus().name)).toEqual(
+      carNames,
+    );
   });
 
-  test('자동차 객체 중 가장 멀리 간 자동차를 선정할 수 있다.', () => {
-    expect(racing.getWinner()).toEqual('efg');
+  test('자동차 객체 중 가장 멀리 간 자동차를 배열에 담아 반환한다..', () => {
+    expect(racing.getWinner()).toEqual(['efg']);
   });
-  test('우승자가 여러 명인 경우 쉼표를 이용해 구분한다', () => {
+  test('우승자가 여러 명인 경우 여러 개의 자동차를 배열에 담아 반환한다.', () => {
     carList.push(new Car('hij', 2));
-    expect(racing.getWinner()).toEqual('efg, hij');
+    expect(racing.getWinner()).toEqual(['efg', 'hij']);
   });
 });
