+++ src/utils/printMessage.js
@@ -1,3 +1,5 @@
+import { FORWARD_DASH, GAME_MESSAGE } from '../constants/systemMessages.js';
+
 const Printer = Object.freeze({
   printMessage(content) {
     console.log(content);
@@ -5,6 +7,17 @@ const Printer = Object.freeze({
 
   printBlank() {
     console.log();
+  },
+
+  printRaceStatus(racing) {
+    racing.carList.forEach((car) => {
+      const { name, position } = car.getCarStatus();
+      Printer.printMessage(`${name} : ${FORWARD_DASH.repeat(position)}`);
+    });
+  },
+
+  printRaceWinner(racing) {
+    Printer.printMessage(`${GAME_MESSAGE.WINNER} ${racing.getWinner()}`);
   },
 });
